# openapi.yaml
openapi: "3.1.0"
info:
  title: GrootAi Data Quality Platform API
  version: "1.0.0"
  description: |
    REST API for GrootAi — an AI-powered data quality governance platform.
    All endpoints are available under both `/api/` (legacy) and `/api/v1/` (current).
  contact:
    name: GrootAi
    url: https://github.com/BalajiHariharan30/GrootAi

servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  - url: https://grootai.onrender.com/api/v1
    description: Production

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Dataset:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        rowCount: { type: integer }
        columns: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }

    Issue:
      type: object
      properties:
        _id: { type: string }
        datasetId: { type: string }
        type: { type: string, enum: [missing_value, format_violation, out_of_range, duplicate, rule_violation] }
        severity: { type: string, enum: [high, medium, low] }
        field: { type: string }
        currentValue: { type: string }
        explanation: { type: string }
        matchConfidence: { type: number }
        status: { type: string, enum: [open, resolved, ignored] }

    RemediationAction:
      type: object
      properties:
        _id: { type: string }
        issueId: { type: string }
        proposedFix: { type: object }
        status: { type: string, enum: [proposed, approved, rejected, rolled_back] }
        aiModel: { type: string }

    Rule:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        description: { type: string }
        condition: { type: object }
        isActive: { type: boolean }

    HealthResponse:
      type: object
      properties:
        status: { type: string }
        platform: { type: string }
        version: { type: string }
        apiVersion: { type: string }
        storageMode: { type: string, enum: [mongodb, in-memory] }
        cacheMode: { type: string, enum: [redis, in-memory] }
        isDemoMode: { type: boolean }

paths:
  /health:
    get:
      summary: Platform health check
      operationId: getHealth
      tags: [System]
      responses:
        "200":
          description: Platform status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }

  /datasets:
    get:
      summary: List all datasets
      operationId: listDatasets
      tags: [Datasets]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Array of datasets }
    post:
      summary: Upload a new dataset (CSV)
      operationId: uploadDataset
      tags: [Datasets]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                name: { type: string }
      responses:
        "201": { description: Dataset created }
        "400": { description: Invalid file or row limit exceeded }

  /datasets/{id}/profile:
    get:
      summary: Get column-level profiling statistics
      operationId: profileDataset
      tags: [Datasets]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Dataset profile }

  /issues:
    get:
      summary: List data quality issues
      operationId: listIssues
      tags: [Issues]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: datasetId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [open, resolved, ignored] }
        - name: severity
          in: query
          schema: { type: string, enum: [high, medium, low] }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200": { description: Paginated issue list }

  /remediation/propose:
    post:
      summary: Propose an AI fix for a single issue
      operationId: proposeFix
      tags: [Remediation]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                issueId: { type: string }
      responses:
        "200": { description: Remediation proposal }

  /remediation/propose-batch:
    post:
      summary: Propose AI fixes for up to 50 issues in one call
      operationId: proposeBatchFix
      tags: [Remediation]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                issueIds:
                  type: array
                  items: { type: string }
                  maxItems: 50
      responses:
        "200": { description: Array of remediation proposals }

  /remediation/{id}/approve:
    post:
      summary: Approve and apply a proposed fix
      operationId: approveFix
      tags: [Remediation]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Fix applied }

  /remediation/{id}/rollback:
    post:
      summary: Roll back an approved fix to original value
      operationId: rollbackFix
      tags: [Remediation]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Fix rolled back }

  /rules:
    get:
      summary: List all rules
      operationId: listRules
      tags: [Rules]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Array of rules }
    post:
      summary: Create a rule (natural language or structured AST)
      operationId: createRule
      tags: [Rules]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                naturalLanguageInput: { type: string }
                condition: { type: object }
      responses:
        "201": { description: Rule created }

  /eval/run:
    post:
      summary: Run the eval benchmark suite
      operationId: runEval
      tags: [Eval]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Eval results }